core: squeeze a little more perf out of block parsing
Most of the time is now spent in input port internals.  To improve
things further we probably have to ditch input ports.

GC time is somewhat high, so maybe that's something else that could be
improved.  I tried a change where I passed around a buffer for
literals instead of creating new bytes all the time, but it didn't
make a difference.